feat(routines): split detail page, default to owner DM, add channel picker
- Decompose RoutineDetailRoute.tsx (was 1717 lines, over the 1500 lint
  budget) into a page shell plus three tab modules and a shared module:
  - RoutineRevisionsTab.tsx — versions list + diff
  - RoutineRunsList.tsx — expandable run rows + detail panel
  - routineDetailShared.tsx — Section, OwnerSummary, Instructions,
    prettyPayload, scheduleSubtext, button style
  - The route file is now 798 lines.

- Default agent-routine fires into the owner's DM instead of #general.
  processAgentJob synthesises the DM channel via the new
  schedulerBroker.EnsureDirectChannel hook when the routine has no
  explicit channel; #general only catches the long tail (routines whose
  owner has no DM yet). Existing routines whose channel was blank get
  the new behaviour on next fire.

- Skip the textual @mention prefix in DMs (it's pure noise in a 1:1)
  while keeping it in shared channels so the assignment stays obvious
  in chat AND survives any downstream text-scanning fallback.

- New "Run in" field in the composer and inline editor. Empty value =
  owner DM; otherwise a team channel slug. The dropdown lists only
  channels the owning agent is actually a member of (filterEligibleChannels
  in RoutineChannelSelect), and excludes DMs so the explicit "Owner DM"
  option stays canonical.

- PATCH /scheduler/{slug} now accepts `channel` as part of the editable
  content shape, snapshotting a revision when it changes.
